# server/src/openapi/openapi_spec.yaml
openapi: 3.0.0
paths:
...
  /api/settings:
    get:
      operationId: SettingsController_getSettings
      parameters: []
      responses:
        '200':
          description: List of all settings
      summary: Get all settings
      tags: &ref_1
        - Settings
    post:
      operationId: SettingsController_updateSettings
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingDto'
      responses:
        '200':
          description: Settings updated successfully
      summary: Update settings
      tags: *ref_1
  /api/settings/radarr:
    get:
      operationId: SettingsController_getRadarrSettings
      parameters: []
      responses:
        '200':
          description: Radarr settings retrieved successfully
      summary: Get Radarr settings
      tags: *ref_1
    post:
      operationId: SettingsController_addRadarrSetting
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RadarrSettingRawDto'
      responses:
        '200':
          description: Radarr setting added successfully
      summary: Add a new Radarr setting
      tags: *ref_1
  /api/settings/sonarr:
    get:
      operationId: SettingsController_getSonarrSettings
      parameters: []
      responses:
        '200':
          description: Sonarr settings retrieved successfully
      summary: Get Sonarr settings
      tags: *ref_1
    post:
      operationId: SettingsController_addSonarrSetting
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SonarrSettingRawDto'
      responses:
        '200':
          description: Sonarr setting added successfully
      summary: Add a new Sonarr setting
      tags: *ref_1
...
  /api/settings/plex/token:
    post:
      operationId: SettingsController_updateAuthToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                plex_auth_token:
                  type: string
                  example: myplexauthtoken
      responses:
        '200':
          description: Plex auth token updated successfully
      summary: Update Plex auth token
      tags: *ref_1
...
  /api/settings/test/radarr:
    post:
      operationId: SettingsController_testRadarr
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RadarrSettingRawDto'
      responses:
        '200':
          description: Radarr settings tested successfully
      summary: Test Radarr settings
      tags: *ref_1
  /api/settings/radarr/{id}:
    put:
      operationId: SettingsController_updateRadarrSetting
      parameters:
        - name: id
          required: true
          in: path
          description: ID of the Radarr setting to update
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RadarrSettingRawDto'
      responses:
        '200':
          description: Radarr setting updated successfully
      summary: Update a Radarr setting
      tags: *ref_1
    delete:
      operationId: SettingsController_deleteRadarrSetting
      parameters:
        - name: id
          required: true
          in: path
          description: ID of the Radarr setting to delete
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Radarr setting deleted successfully
      summary: Delete a Radarr setting
      tags: *ref_1
  /api/settings/test/sonarr:
    post:
      operationId: SettingsController_testSonarr
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SonarrSettingRawDto'
      responses:
        '200':
          description: Sonarr settings tested successfully
      summary: Test Sonarr settings
      tags: *ref_1
  /api/settings/sonarr/{id}:
    put:
      operationId: SettingsController_updateSonarrSetting
      parameters:
        - name: id
          required: true
          in: path
          description: ID of the Sonarr setting to update
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SonarrSettingRawDto'
      responses:
        '200':
          description: Sonarr setting updated successfully
      summary: Update a Sonarr setting
      tags: *ref_1
    delete:
      operationId: SettingsController_deleteSonarrSetting
      parameters:
        - name: id
          required: true
          in: path
          description: ID of the Sonarr setting to delete
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Sonarr setting deleted successfully
      summary: Delete a Sonarr setting
      tags: *ref_1
...
  /api/settings/cron/validate:
    post:
      operationId: SettingsController_validateSingleCron
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronScheduleDto'
      responses:
        '200':
          description: CRON expression validated successfully
      summary: Validate a CRON expression
      tags: *ref_1
...
info:
  title: Maintainerr API
  description: Documentation for Maintainerr API
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    SettingDto:
      type: object
      properties:
        id:
          type: number
          description: Unique identifier for the setting
          example: 1
        clientId:
          type: string
          description: Client ID associated with the setting
          example: 12345-abcde
        applicationTitle:
          type: string
          description: Title of the application
          example: My App
        applicationUrl:
          type: string
          description: Base URL of the application
          example: https://myapp.example.com
        apikey:
          type: string
          description: API key for accessing the application
          example: abc123xyz456
        locale:
          type: string
          description: Locale setting for the application
          example: en
        cacheImages:
          type: number
          description: Cache images setting (1 = enabled, 0 = disabled)
          example: 1
        plex_name:
          type: string
          description: Name of the Plex server
          example: My Plex Server
        plex_hostname:
          type: string
          description: Hostname of the Plex server
          example: plex.example.com
        plex_port:
          type: number
          description: Port number for the Plex server
          example: 32400
        plex_ssl:
          type: number
          description: Whether Plex server uses SSL (1 = enabled, 0 = disabled)
          example: 1
        plex_auth_token:
          type: string
          description: Authentication token for the Plex server
          example: plex_auth_token_example
        overseerr_url:
          type: string
          description: URL of the Overseerr instance
          example: https://overseerr.example.com
        overseerr_api_key:
          type: string
          description: API key for the Overseerr instance
          example: overseerr_api_key_example
        tautulli_url:
          type: string
          description: URL of the Tautulli instance
          example: https://tautulli.example.com
        tautulli_api_key:
          type: string
          description: API key for the Tautulli instance
          example: tautulli_api_key_example
        collection_handler_job_cron:
          type: string
          description: CRON expression for the collection handler job
          example: 0 0 * * *
        rules_handler_job_cron:
          type: string
          description: CRON expression for the rules handler job
          example: 0 30 * * *
      required:
        - id
        - clientId
        - applicationTitle
        - applicationUrl
        - apikey
        - locale
        - cacheImages
        - plex_name
        - plex_hostname
        - plex_port
        - plex_ssl
        - plex_auth_token
        - overseerr_url
        - overseerr_api_key
        - tautulli_url
        - tautulli_api_key
        - collection_handler_job_cron
        - rules_handler_job_cron
    RadarrSettingRawDto:
      type: object
      properties:
        serverName:
          type: string
        url:
          type: string
        apiKey:
          type: string
      required:
        - serverName
        - url
        - apiKey
    SonarrSettingRawDto:
      type: object
      properties:
        serverName:
          type: string
        url:
          type: string
        apiKey:
          type: string
      required:
        - serverName
        - url
        - apiKey
    CronScheduleDto:
      type: object
      properties: {}
...